update controller and service class
